feat(23618): Create experiments architecture

+++ platform-sdk/consensus-otter-tests/src/testPerformance/java/org/hiero/otter/test/performance/benchmark/ConsensusLayerBenchmark.java
@@ -162,7 +162,7 @@ public class ConsensusLayerBenchmark {
         parseFromLogs(network.newLogResults(), BenchmarkServiceLogParser::parseMeasurement, collector::addEntry);
         // Make sure the benchmark run is valid
         assertEquals(
-                params.transactionCount() * params.numberOfNodes(),
+                (long) params.transactionCount() * params.numberOfNodes(),
                 collector.computeStatistics().totalMeasurements(),
                 "The benchmark is invalid as some of the transactions sent were not measured");
         final String report = collector.generateReport();
